templates/requirements.txt
flask==2.0.1
faiss-cpu==1.7.2
numpy==1.21.2
sentence-transformers==2.2.2
openai==1.0.0